+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 # 4.1.6
 
-* Add multi-attach support to EBS volume
+* Change EBS volume type to gp3
 * Add DeletionPolicy: Snapshot to EBS volume
+* Add UpdateReplacePolicy: Snapshot to EBS volume
 
 # 4.1.5