+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change EBS volume type to gp3
 * Add DeletionPolicy: Snapshot to EBS volume
 * Add UpdateReplacePolicy: Snapshot to EBS volume
+* Encrypt EBS volume
 
 # 4.1.5